pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polyapp-llm-yaml-generator"
version = "0.1.0"
description = "A project to generate YAML configuration files using LLM and Pydantic models"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "pydantic",
    "openai",
    "python-dotenv",
    "pyyaml"
]

[tool.setuptools]
packages = ["generate_llm_yaml"]

[tool.ruff]
exclude = ["scripts/"]
line-length = 88
target-version = "py313"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = ["E501", "F841", "UP015", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.isort]
profile = "black"
line_length = 88 